Introducing virtio-increment-edu.
The main goal is to create an educational example to be used as template
or guideline for contributors when they wish to create a new virtio
device and to document "the right way" to do so.

The device functionality is simple as possible so contributors can have
their focus on the virtio protocol rather on the device functionality to
give them an "easy start".

It consists of several parts:

    1. The device specification
        * this patch content

    2. The device implementation for Qemu-KVM hypervisor
        * it will hopefully be added to the Qemu project
        * for now it can be found at
https://github.com/ybettan/qemu/blob/\
                virtio/hw/virtio/virtio-example.c

    3. The device driver for linux
        * it will hopefully be added to linux
        * for now it can be found at https://github.com/ybettan/\
                QemuDeviceDrivers/blob/master/virtio/virtio_example_driver.c

    4. A blog on virtio
        * introducing the virtio concept
        * gives some motivation for virtio-devices to be used
        * bring extra documentation on "how to write":
            - device specification
            - device implementation
            - device driver for linux
        * it can be found at https://howtovms.wordpress.com

Signed-off-by: Yoni Bettan <ybettan@redhat.com>